unittest: migrate unittests to test dir, run under pytest
