+++ lib/python/rose/tests/test_config.py
